--- a/cli/.gitignore
+++ b/cli/.gitignore
@@ -4,3 +4,4 @@ dbdata/
 libcoredb/
 cli/.yalc/@ledgerhq/live-common/
 cli/yalc.lock
+src/commands-index.js