--- a/cli/scripts/gen.sh
+++ b/cli/scripts/gen.sh
@@ -1,0 +1,23 @@
+#!/bin/bash
+
+set -e
+cd $(dirname $0)
+cd ../src
+
+gen() {
+  for d in commands/*.js; do
+    path=${d%.*}
+    name=${path#*/}
+    echo 'import '$name' from "./'$path'";'
+  done
+  echo ''
+  echo 'export default {'
+  for d in commands/*.js; do
+    path=${d%.*}
+    name=${path#*/}
+    echo '  '$name','
+  done
+  echo '};'
+}
+
+gen > commands-index.js