--- a/cli/src/commands-index.js
+++ b/cli/src/commands-index.js
@@ -1,0 +1,57 @@
+import app from "./commands/app";
+import appsUpdateTestAll from "./commands/appsUpdateTestAll";
+import balanceHistory from "./commands/balanceHistory";
+import broadcast from "./commands/broadcast";
+import deviceAppVersion from "./commands/deviceAppVersion";
+import deviceInfo from "./commands/deviceInfo";
+import deviceVersion from "./commands/deviceVersion";
+import discoverDevices from "./commands/discoverDevices";
+import exportAccounts from "./commands/exportAccounts";
+import firmwareRepair from "./commands/firmwareRepair";
+import firmwareUpdate from "./commands/firmwareUpdate";
+import genuineCheck from "./commands/genuineCheck";
+import getAccountNetworkInfo from "./commands/getAccountNetworkInfo";
+import getAddress from "./commands/getAddress";
+import getTransactionStatus from "./commands/getTransactionStatus";
+import libcoreReset from "./commands/libcoreReset";
+import libcoreSetPassword from "./commands/libcoreSetPassword";
+import liveData from "./commands/liveData";
+import managerListApps from "./commands/managerListApps";
+import proxy from "./commands/proxy";
+import receive from "./commands/receive";
+import repl from "./commands/repl";
+import send from "./commands/send";
+import signMessage from "./commands/signMessage";
+import sync from "./commands/sync";
+import validRecipient from "./commands/validRecipient";
+import version from "./commands/version";
+
+export default {
+  app,
+  appsUpdateTestAll,
+  balanceHistory,
+  broadcast,
+  deviceAppVersion,
+  deviceInfo,
+  deviceVersion,
+  discoverDevices,
+  exportAccounts,
+  firmwareRepair,
+  firmwareUpdate,
+  genuineCheck,
+  getAccountNetworkInfo,
+  getAddress,
+  getTransactionStatus,
+  libcoreReset,
+  libcoreSetPassword,
+  liveData,
+  managerListApps,
+  proxy,
+  receive,
+  repl,
+  send,
+  signMessage,
+  sync,
+  validRecipient,
+  version,
+};