--- a/cli/src/commands/balanceHistory.js
+++ b/cli/src/commands/balanceHistory.js
@@ -1,0 +1,107 @@
+// @flow
+
+import { BigNumber } from "bignumber.js";
+import asciichart from "asciichart";
+import invariant from "invariant";
+import { map } from "rxjs/operators";
+import { toBalanceHistoryRaw } from "@ledgerhq/live-common/lib/account";
+import type { PortfolioRange } from "@ledgerhq/live-common/lib/types";
+import {
+  getBalanceHistory,
+  getRanges
+} from "@ledgerhq/live-common/lib/portfolio";
+import { formatCurrencyUnit } from "@ledgerhq/live-common/lib/currencies";
+import { scan, scanCommonOpts } from "../scan";
+import type { ScanCommonOpts } from "../scan";
+
+const histoFormatters = {
+  default: (histo, account) =>
+    histo
+      .map(
+        ({ date, value }) =>
+          date.toISOString() +
+          " " +
+          formatCurrencyUnit(account.unit, value, {
+            showCode: true,
+            disableRounding: true
+          })
+      )
+      .join("\n"),
+
+  json: histo => toBalanceHistoryRaw(histo),
+
+  asciichart: (history, account) =>
+    "\n" +
+    "".padStart(22) +
+    account.name +
+    ": " +
+    formatCurrencyUnit(account.unit, account.balance, {
+      showCode: true,
+      disableRounding: true
+    }) +
+    "\n" +
+    asciichart.plot(
+      history.map(h =>
+        h.value.div(BigNumber(10).pow(account.unit.magnitude)).toNumber()
+      ),
+      {
+        height: 10,
+        format: value =>
+          formatCurrencyUnit(
+            account.unit,
+            BigNumber(value).times(BigNumber(10).pow(account.unit.magnitude)),
+            {
+              showCode: true,
+              disableRounding: true
+            }
+          ).padStart(20)
+      }
+    )
+};
+
+function asPortfolioRange(period: string): PortfolioRange {
+  const ranges = getRanges();
+  invariant(
+    ranges.includes(period),
+    "invalid period. valid values are %s",
+    ranges.join(" | ")
+  );
+  // $FlowFixMe
+  return period;
+}
+
+export default {
+  description: "Get the balance history for accounts",
+  args: [
+    ...scanCommonOpts,
+    {
+      name: "period",
+      alias: "p",
+      type: String,
+      desc: getRanges().join(" | ")
+    },
+    {
+      name: "format",
+      alias: "f",
+      type: String,
+      typeDesc: Object.keys(histoFormatters).join(" | "),
+      desc: "how to display the data"
+    }
+  ],
+  job: (
+    opts: ScanCommonOpts & {
+      format: string,
+      period: string
+    }
+  ) =>
+    scan(opts).pipe(
+      map(account => {
+        const histo = getBalanceHistory(
+          account,
+          asPortfolioRange(opts.period || "month")
+        );
+        const format = histoFormatters[opts.format || "default"];
+        return format(histo, account);
+      })
+    )
+};