--- a/cli/src/commands/deviceAppVersion.js
+++ b/cli/src/commands/deviceAppVersion.js
@@ -1,0 +1,12 @@
+// @flow
+
+import { from } from "rxjs";
+import { withDevice } from "@ledgerhq/live-common/lib/hw/deviceAccess";
+import getAppAndVersion from "@ledgerhq/live-common/lib/hw/getAppAndVersion";
+import { deviceOpt } from "../scan";
+
+export default {
+  args: [deviceOpt],
+  job: ({ device }: $Shape<{ device: string }>) =>
+    withDevice(device || "")(t => from(getAppAndVersion(t)))
+};