--- a/cli/src/commands/deviceInfo.js
+++ b/cli/src/commands/deviceInfo.js
@@ -1,0 +1,12 @@
+// @flow
+
+import { from } from "rxjs";
+import { withDevice } from "@ledgerhq/live-common/lib/hw/deviceAccess";
+import getDeviceInfo from "@ledgerhq/live-common/lib/hw/getDeviceInfo";
+import { deviceOpt } from "../scan";
+
+export default {
+  args: [deviceOpt],
+  job: ({ device }: $Shape<{ device: string }>) =>
+    withDevice(device || "")(t => from(getDeviceInfo(t)))
+};