--- a/cli/src/commands/deviceVersion.js
+++ b/cli/src/commands/deviceVersion.js
@@ -1,0 +1,12 @@
+// @flow
+
+import { from } from "rxjs";
+import { withDevice } from "@ledgerhq/live-common/lib/hw/deviceAccess";
+import getVersion from "@ledgerhq/live-common/lib/hw/getVersion";
+import { deviceOpt } from "../scan";
+
+export default {
+  args: [deviceOpt],
+  job: ({ device }: $Shape<{ device: string }>) =>
+    withDevice(device || "")(t => from(getVersion(t)))
+};