--- a/cli/src/commands/libcoreReset.js
+++ b/cli/src/commands/libcoreReset.js
@@ -1,0 +1,10 @@
+// @flow
+import { withLibcore } from "@ledgerhq/live-common/lib/libcore/access";
+
+export default {
+  args: [],
+  job: () =>
+    withLibcore(async core => {
+      await core.getPoolInstance().freshResetAll();
+    })
+};