--- a/cli/src/qr.js
+++ b/cli/src/qr.js
@@ -1,0 +1,12 @@
+// @flow
+
+import { Observable } from "rxjs";
+import qrcode from "qrcode-terminal";
+
+export const asQR = (str: string) =>
+  Observable.create(o =>
+    qrcode.generate(str, r => {
+      o.next(r);
+      o.complete();
+    })
+  );