Remaining:

- [ ] Integrate in mobile. and test it (with help of QA)
- [ ] Integrate in LLD v2.
- [ ] Need to test in a minimal PoC that pagination seems to work. (low effort on this because we want to prove the contract is good, not that the implementation is 100% working as we don't move to it yet)

---

## Performance of ETH scanning (when you have ERC20)

use ERC20Balances to fetch in batch instead of loop calls. #439 

## Performance of second synchro of heavy accounts

**Heavy accounts:** an account that many operations and/or many "outputs" in receiving operations (typically when pool mining).

Using a testing account that have 100 operations where each operation tx have 6000 outputs,
we can reduce the sync time from 20 seconds to 3 seconds so it's about a 10x perf gain for this context. But this only concerns the second sync.

### tradeoff for heavy inputs/outputs

A tradeoff is to introduce an environment, currently set to 100, to limit the maximum number of inputs/outputs that will appear in Ledger Live for better performance.

## Preparing terrain of pagination

We have changes a bit the typing of Account, AccountBridge, CurrencyBridge to prepare future needs including future support of paginations on operation, especially in context of PoS coins like Tezos.

The general idea is that, as a (lib) user, I can do

```
sync(account, { paginationConfig: { operations: 40 } })
```

scroll a bit and load more ahead of time:

```
sync(account, { paginationConfig: { operations: 80 } })
...
```

the goal is NOT to decrement the pagination config, but always "push it forward".
The UI will have to manage such state, but not persist it.

### `Account#operationsCount`

We introduce `operationsCount` on Account and Account likes type.

```diff
+  operationsCount: number,
```

The RawAccount versions will have it optional so we can be backward compatible on the data.

```diff
+  operationsCount?: number,
```

### `Account#balanceHistory` #429 

We also introduces the "balanceHistory" field on Account and Account likes type.
This allows to cache in the Account model itself the historical values of the account balance instead of calculating at runtime in JavaScript.
This is actually necessary for pagination.

**On Account and Account likes:**

```diff
+ balanceHistory?: BalanceHistoryMap
```

**On AccountRaw and AccountRaw likes:**

```diff
+ balanceHistory?: BalanceHistoryRawMap
```

```js
export type BalanceHistory = Array<{ date: Date, value: BigNumber }>;
export type BalanceHistoryMap = {
  [_: PortfolioRange]: BalanceHistory
};

// Raw equivalent, so it can also be persisted:

export type BalanceHistoryRaw = Array<[string, string]>;
export type BalanceHistoryRawMap = {
  [_: PortfolioRange]: BalanceHistoryRaw
};
```

> The goal is we are going to incrementally implement this feature and consistently with the support of not of pagination. The fact pagination is either lazy or not is a implementation specific problem and does not impact at all the "user land" side. So we can always implement the bridge like before (no pagination, no balance history), and later implement pagination (but when we do, then balance history is necessary).

#### CLI support

```
$ ledger-live balanceHistory --format asciichart -p month

                      Tezos 1 (galleon): XTZ 17.15072
        XTZ 17.15072 ┤  ╭──────────────────────────
       XTZ 15.848349 ┤  │
       XTZ 14.545978 ┤  │
       XTZ 13.243607 ┤  │
       XTZ 11.941236 ┤  │
       XTZ 10.638866 ┤  │
        XTZ 9.336495 ┤  │
        XTZ 8.034124 ┤  │
        XTZ 6.731753 ┼─╮│
        XTZ 5.429382 ┤ ││
        XTZ 4.127011 ┤ ╰╯

                      Tezos 2 (galleon): XTZ 0.048063
        XTZ 0.078063 ┼─╮
        XTZ 0.075063 ┤ │
        XTZ 0.072063 ┤ │
        XTZ 0.069063 ┤ │
        XTZ 0.066063 ┤ │
        XTZ 0.063063 ┤ │
        XTZ 0.060063 ┤ │
        XTZ 0.057063 ┤ │
        XTZ 0.054063 ┤ │
        XTZ 0.051063 ┤ │
        XTZ 0.048063 ┤ ╰───────────────────────────

                      Tezos 3 (galleon): XTZ 2.914587
        XTZ 2.914587 ┤                           ╭─
         XTZ 2.91434 ┤                           │
        XTZ 2.914092 ┤                           │
        XTZ 2.913845 ┤                           │
        XTZ 2.913597 ┤                           │
         XTZ 2.91335 ┤                           │
        XTZ 2.913102 ┤                           │
        XTZ 2.912855 ┤                           │
        XTZ 2.912607 ┤                           │
         XTZ 2.91236 ┤                           │
        XTZ 2.912112 ┼───────────────────────────╯

                      Tezos 4 (galleon): XTZ 0
               XTZ 0 ┼─────────────────────────────
```

### SyncConfig and PaginationConfig concepts

We introduce a contextual parameter to all account synchronisations.

```js
// A pagination config holds the user's pagination state
// this is a state that usually should leave during the app lifecycle, but is not persisted
// it drives the number of operations to poll in accounts
// when a user paginate more, the number should accordingly be incremented
// The UI should manage scrolling ahead of time (e.g. if 30 ops is displayed and UI have pages of 20 ops, the UI can already request to poll 70 ops so it have 2 pages in advance)
// The UI must always do max() to keep the increasing the counter and not going back to lower value: that optim the sync to not recompute things too much
export type PaginationConfig = {
  // operations to pull for each account
  operationsPerAccountId?: {
    [_: string]: number
  },
  // if define and there is no specific account in operationsPerAccountId,
  // this will be the operations count used
  operations?: number
};

export type SyncConfig = {
  paginationConfig: PaginationConfig,
  // allows to disable the synchronization part – typically to only paginate more
  withoutSynchronize?: boolean
};
```

### AccountBridge: `startSync` becomes `sync`

To establish a clear signature changes, we rename `startSync` into `sync`

```diff
-  startSync(
+  sync(
     initialAccount: Account,
-    observation: boolean
+    syncConfig: SyncConfig
   ): Observable<(Account) => Account>;
```

We have the syncConfig parameter here that could be used for more things in the future than the pagination.

The pagination is currently implemented already but will not be connected to the UI yet.

### CurrencyBridge: `scanAccountsOnDevice` becomes `scanAccounts`

```diff
-  scanAccountsOnDevice(
+  scanAccounts({
     currency: CryptoCurrency,
     deviceId: DeviceId,
-    scheme?: ?DerivationMode
-  ): Observable<ScanAccountEvent>;
+    scheme?: ?DerivationMode,
+    syncConfig: SyncConfig
+  }): Observable<ScanAccountEvent>;
```

### CLI to support paginationOperations

```sh
$ ledger-live sync --file out.json --paginateOperations 10
# ... 10 operations displayed ...
$ ledger-live sync --file out.json --paginateOperations 20
# ... 20 operations displayed ...
```

## `signAndBroadcast` is replaced by `signOperation` and `broadcast`

We decided to split `signAndBroadcast` in two parts to allow more powerful features in the future like being able to first sign with a device and only broadcast later. Similarly, we might want to display the signature to user before he validates.

This makes the contract more clear:

```diff
-  // finalizes the transaction by
-  // - signing it with the ledger device
-  // - broadcasting it to network
-  // - retrieve and return the optimistic Operation that this transaction is likely to create in the future
-  signAndBroadcast(
+  // finalizing a transaction by signing it with the ledger device
+  // This results of a "signed" event with a signedOperation
+  // than can be locally saved and later broadcasted
+  signOperation({
     account: Account,
     transaction: T,
     deviceId: DeviceId
-  ): Observable<SignAndBroadcastEvent>;
+  }): Observable<SignOperationEvent>;
+
+  // broadcasting a signed transaction to network
+  // returns an optimistic Operation that this transaction is likely to create in the future
+  broadcast({
+    account: Account,
+    signedOperation: SignedOperation
+  }): Promise<Operation>;
```

### `SignedOperation`

```js
export type SignedOperation = {|
  // prepared version of Operation before it's even broadcasted
  // .id/.hash is potentially not settled yet
  operation: Operation,
  // device signature hex
  signature: string,
  // date calculated as expiring
  expirationDate: ?Date
|};
```

### `SignOperationEvent`

We also take the opportunity to add new events in `SignOperationEvent` for better UX:

```js
export type SignOperationEvent =
  // Used when lot of exchange is needed with the device to visually express a progress
  // It can be used before and/or after the signature
  // only used if it can takes >1s to show a visual progress to user (typically UTXO streaming)
  | { type: "device-streaming", progress: number } // optional
  // REQUIRED Indicates that a signature is now appearing and awaited on the device to confirm
  | { type: "device-signature-requested" }
  // REQUIRED Indicates user have confirmed the transaction
  | { type: "device-signature-granted" }
  // REQUIRED payload of the resulting signed operation
  | { type: "signed", signedOperation: SignedOperation };
```

Example:

![](https://user-images.githubusercontent.com/211411/71770088-30f8d580-2f29-11ea-94ff-14f308be7aa2.png)

### CLI support

```sh
$ ledger-live send --scheme native_segwit -i 0 --self-transaction --disable-broadcast | jq 'select(.type=="signed") | .signedOperation' > tx.json

$ cat tx.json
{"operation":{...},"signature":"010000000001038949782c6...ae73ef00000000","expirationDate":null}

$ ledger-live broadcast --signed-operation tx.json --scheme native_segwit -i 0
{"id":"libcore:1:bitcoin:...:native_segwit-...-OUT","hash":"...","type":"OUT","blockHash":null,"blockHeight":null,"senders":["bc1q..78u","bc1q..rqg","bc1q..hp"],"recipients":["bc1q..jg","bc1..ka"],"accountId":"libcore:1:bitcoin:..:native_segwit","extra":{},"date":"2020-01-04T11:00:00.000Z","value":"100282","fee":"282"}
```

## Tradeoffs

**hourly rate and hourly granularity is not yet enabled**

but it's in progress.

- we need a new libcore version
- we need to have a more performance Countervalues API

**pagination will not be implemented (on UI side) yet**

but we should plan to make it possible.

**Tezos will need some evolution to enable pagination**

- Find a solution for Tezos delegation status: we rely on fact we have all ops... can we query with libcore with a filter on the delegation kinds?
  - How to we materialize that in our model (we still want to paginate ops). and idea would be to have "lastDelegationOp" but that feels too specific (what about TRON? won't we have more usecase of "analyzing data of ops to get an info"?)


## Benchmarks (without testing pagination)

There is a new folder "benchmark", that runs some tests on different version of live-common / libcore to compare performance. Here is the current result of this PR:

```
yarn start runTests -n 10

boot time (10 runs, 3 targets) – (total 60867ms)
  current: avg 1990ms median 1797ms
  latest: median 1818ms (1% slower) (avg 2030ms)
  v10: median 1791ms (similar) (avg 2067ms)

first sync bitcoin (10 runs, 3 targets) – (total 221867ms)
  current: avg 7385ms median 7385ms
  latest: median 7268ms (2% faster) (avg 7345ms)
  v10: median 7414ms (similar) (avg 7409ms)

second sync bitcoin (10 runs, 3 targets) – (total 93304ms)
  current: avg 3184ms median 3013ms
  latest: median 3105ms (3% slower) (avg 3050ms)
  v10: median 3009ms (similar) (avg 3096ms)

first sync ethereum (10 runs, 3 targets) – (total 288877ms)
  current: avg 4639ms median 3569ms
  latest: median 10405ms (192% slower) (avg 11619ms)
  v10: median 13410ms (276% slower) (avg 12577ms)

second sync ethereum (10 runs, 3 targets) – (total 258592ms)
  current: avg 3789ms median 3270ms
  latest: median 9792ms (199% slower) (avg 11423ms)
  v10: median 10333ms (216% slower) (avg 10647ms)
```